pyproject.toml
[tool.poetry]
name = "lance-benchmark"
version = "0.1.0"
description = "A time series benchmark for the lance file format"
authors = ["Nico Bautista Hobin <nico@calligram.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.1"
pandas = "^2.1.1"
pyarrow = "^13.0.0"
alpaca-py = "^0.12.0"
duckdb = "^0.9.1"
pylance = "^0.8.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.10.0"
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
